+++ src/javascript/api/updater.js
@@ -17,9 +17,12 @@ module.exports = {
     const ghrepo = client.repo('Graviton-Code-Editor/Graviton-App')
     ghrepo.releases(function (err, res, body) {
       if (!err) {
+        console.log(res,body);
         for (i = 0; i < res.length + 1; i++) {
           if (i < res.length + 1) {
-            if (res[i].tag_name > GravitonInfo.version) {
+            console.log(i,res.length);
+            console.log(res[i].tag_name,GravitonInfo.version)
+            if (semver.gt(res[i].tag_name,GravitonInfo.version)) {
               // New update detected
               new Dialog({
                 id: 'update',